+++ app/components/address-mapper.tsx
@@ -12,6 +12,7 @@ const addressLookupCoordinates = {
   geometry: { type: "Point", coordinates: [-122.408020683, 37.801698301] },
 };
 const defaultCoords = addressLookupCoordinates.geometry.coordinates ?? [];
+const toggledStatesDefaults = [true, true, true];
 
 interface AddressMapperProps {
   softStoryData: FeatureCollection<Geometry>;
@@ -44,6 +45,9 @@ const AddressMapper: React.FC<AddressMapperProps> = ({
   const [searchedAddress, setSearchedAddress] = useState("");
   const [addressHazardData, setAddressHazardData] = useState<object>({});
   const [isHazardDataLoading, setHazardDataLoading] = useState(false);
+  const [toggledStates, setToggledStates] = useState<boolean[]>(
+    toggledStatesDefaults
+  );
   const [layerToggleObj, setLayerToggleObj] = useState<LayerToggleObjProps>({
     layerId: "",
     toggleState: true,
@@ -98,6 +102,8 @@ const AddressMapper: React.FC<AddressMapperProps> = ({
             <ReportHazards
               addressHazardData={addressHazardData}
               isHazardDataLoading={isHazardDataLoading}
+              toggledStates={toggledStates}
+              setToggledStates={setToggledStates}
               setLayerToggleObj={setLayerToggleObj}
             />
           </Box>
